+++ archer/src/materials/normal.rs
@@ -7,7 +7,7 @@ use crate::materials::base::Material;
 pub struct NormalMaterial {}
 
 impl Material for NormalMaterial {
-    fn bounce(&self, incoming: &Ray, hit_point: Point3<f32>, normal: Normal<f32>) -> Ray {
+    fn bounce(&self, incoming: &Ray, hit_point: Point3<f64>, normal: Normal<f64>) -> Ray {
         let _ = (incoming, hit_point, normal);
         return Ray::new_empty();
     }
@@ -15,10 +15,10 @@ impl Material for NormalMaterial {
     fn add_color(
         &self,
         incoming: &Ray,
-        hit_point: Point3<f32>,
-        normal: Vector3<f32>,
-        color: &Color<f32>
-    ) -> Point3<f32> {
+        hit_point: Point3<f64>,
+        normal: Vector3<f64>,
+        color: &Color<f64>
+    ) -> Point3<f64> {
         let _ = (incoming, hit_point, color);
         return Color::new(
             0.5 * (normal.x + 1.0) * 255.0,
