+++ archer/src/materials/solid.rs
@@ -5,11 +5,11 @@ use crate::vectors::{Color, Normal};
 use crate::materials::base::Material;
 
 pub struct SolidColor {
-    pub color: Color<f32>,
+    pub color: Color<f64>,
 }
 
 impl Material for SolidColor {
-    fn bounce(&self, incoming: &Ray, hit_point: Point3<f32>, normal: Normal<f32>) -> Ray {
+    fn bounce(&self, incoming: &Ray, hit_point: Point3<f64>, normal: Normal<f64>) -> Ray {
         let _ = (incoming, hit_point, normal);
         return Ray::new_empty();
     }
@@ -17,10 +17,10 @@ impl Material for SolidColor {
     fn add_color(
         &self,
         incoming: &Ray,
-        hit_point: Point3<f32>,
-        normal: Vector3<f32>,
-        color: &Color<f32>
-    ) -> Point3<f32> {
+        hit_point: Point3<f64>,
+        normal: Vector3<f64>,
+        color: &Color<f64>
+    ) -> Point3<f64> {
         let _ = (incoming, hit_point, normal, color);
         return self.color;
     }
