+++ archer/src/ray.rs
@@ -4,23 +4,23 @@ use nalgebra::{Point3, Vector3};
 // Ray struct
 #[derive(Debug, Clone, Copy)]
 pub struct Ray {
-    pub origin: Point3<f32>,
-    pub direction: Vector3<f32>,
+    pub origin: Point3<f64>,
+    pub direction: Vector3<f64>,
 }
 
 impl Ray {
     pub fn new_empty() -> Ray {
         return Ray {
-            origin: Point3::new(f32::NAN, f32::NAN, f32::NAN),
-            direction: Vector3::new(f32::NAN, f32::NAN, f32::NAN),
+            origin: Point3::new(f64::NAN, f64::NAN, f64::NAN),
+            direction: Vector3::new(f64::NAN, f64::NAN, f64::NAN),
         };
     }
 
-    pub fn position_at(self, t: f32) -> Point3<f32> {
+    pub fn position_at(self, t: f64) -> Point3<f64> {
         return self.origin + t * self.direction;
     }
 
-    pub fn to_bvh_ray(self) -> BvhRay<f32, 3> {
+    pub fn to_bvh_ray(self) -> BvhRay<f64, 3> {
         return BvhRay::new(self.origin, self.direction);
     }
 }
